requirements.txt
databricks-connect>=15.4.0
databricks-sdk>=0.76.0
databricks-feature-engineering>=0.14.0
mlflow>=3.1.0
lightgbm>=4.5.0
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
shap>=0.46.0
pytest>=8.3.0
pytest-cov>=6.0.0
ruff>=0.8.0
